drivers: ethernet: Fix assert on receiving large PDUs
This patch fixes an assert in net_buf_simple_add() function when the
lan_9250 driver would add the sample pointer to the net buffer library
on receiving large (>128 bytes) packat. This fix was to use the frags
field of the net_buf structure when adding fragmented packets to the net
buffer.

Tested with a nRF54L15 DK and eth3 click board (with nordic connect sdk
2.8.0 and this patch).

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@autostoresystem.com>